requirements.txt
python-telegram-bot==20.7
requests==2.32.5
openai==2.7.1